create a onTypeRange formatting provider, separate multilineString and Indent on paste providers
